Move the calibration code to a subdirectory
